feat(react-ui-dashboards): inline view↔edit toggle + wired hover toolbar
Merge the view and edit surfaces behind an inline toggle (Grafana/luzmo style):
the dashboard view page now flips to edit mode in place — no navigation — and
drops back to the rendered view on exit (post-save invalidation refreshes it).

- Extract the composer body into a shared <DashboardComposer dashboardId onExit>
  reused by DashboardEditPage (route preserved via rootSlot='dashboard-edit-page')
  and mounted inline by DashboardViewPage behind an isEditing toggle.
- Wire the editor hover toolbar: Edit → open WidgetConfigDrawer, Duplicate →
  duplicateWidget, Delete → ConfirmActionDialog + removeWidget. Toolbar added to
  the cell-click delegation ignore list.
- Export DashboardComposer / DashboardComposerProps.

Read renderer stays position-based (the render bundle DashboardRenderedWidget
carries no x/y) — coordinate-aware rendering is a follow-up needing a backend
render-DTO change.

Refs #840
